Adjust numeric IDs

+++ src/main/java/com/mercadopago/client/payment/PaymentCreateRequest.java
@@ -29,7 +29,7 @@ public class PaymentCreateRequest {
   private final String callbackUrl;
 
   /** Discount campaign ID. */
-  private final Integer campaignId;
+  private final Long campaignId;
 
   /** Determines if the payment should be captured (true) or just reserved (false). */
   private final Boolean capture;
@@ -47,7 +47,7 @@ public class PaymentCreateRequest {
   private final String description;
 
   /** Id of the scheme for the absorption of financing fee. */
-  private final Integer differentialPricingId;
+  private final Long differentialPricingId;
 
   /** ID given by the merchant in their system. */
   private final String externalReference;
